Solved the irregular grid lines problem.
1. Removed last internal use of CGFloat from PixelGraphics.
2. Moved Grid specific functions from OCR to Grid.
3. Simplified the xRuns and yRuns arrays from [[PGOLine]] to [PGOLine].
4. Used the trim function to eliminate grid line debris from the edges
   of the Grid cells.

+++ Sudoku/SudokuDocument/OCR/PixelGraphics.swift
@@ -144,9 +144,9 @@ struct PGOLine {
         self.y = start ... end
     }
     
-    var length: CGFloat {
-        if x.lowerBound == x.upperBound { return CGFloat( y.upperBound - y.lowerBound + 1 ) }
-        return CGFloat( x.upperBound - x.lowerBound + 1 )
+    var length: Int {
+        if x.lowerBound == x.upperBound { return y.upperBound - y.lowerBound + 1 }
+        return x.upperBound - x.lowerBound + 1
     }
 }
